refactor: import enums directly from model/enums, drop re-exports
Remove the compatibility re-exports added in the previous commit
(model/types.ts `export * from './enums'`, the event-enum re-export in
engine/events.ts, and EffectType in effect-types.ts). Every consumer now
imports its enums straight from model/enums; the package barrel exports
model/enums so external consumers (game-ui) still resolve via
@mtg-utils/engine-core.

Build, lint, format, and the full vitest suite pass.

[CLAUDE]
