+++ libs/engine-core/src/lib/engine/validators/concede.validator.ts
@@ -1,9 +1,9 @@
 import { ok, type Result } from '@mtg-utils/engine-util';
 
 import type { Concede } from '../../actions/action';
+import { GameEventType, PlayerLostReason } from '../../model/enums';
 import type { GameState } from '../../model/game-state';
 import type { GameEvent } from '../events';
-import { GameEventType, PlayerLostReason } from '../events';
 
 export const validateConcede = (_state: GameState, action: Concede): Result<GameEvent[], string> =>
   ok<GameEvent[]>([
